pyproject.toml
[tool.poetry]
name = "pyproject-toml"
version = "0.1.0"
description = ""
authors = ["tjazerzen <tjaz.erzen@zerodays.dev>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
jwt = "^1.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
